Move enum persistence test to jpa-h2 integration test for native coverage
The deployment-level test doesn't verify enum reflection works in native
mode. Move it to the jpa-h2 integration test which runs in both JVM and
native mode via JPAFunctionalityInGraalITCase.
