--- extensions/hibernate-orm/deployment/src/test/java/io/quarkus/hibernate/orm/enums/Status.java
@@ -1,6 +1,0 @@
-package io.quarkus.hibernate.orm.enums;
-
-public enum Status {
-    LIVING,
-    DECEASED
-}
